Moved the ProjectEnvironmentsRequested handler to tasks. Made task running adding fluent.
